// app/stacks/types.ts
export type PhonebookStackParamList = {
	PhoneBook: undefined;
	Contact: {
		uid: string;
		name: string;
		phone: string;
	};
	RoomView:
		| {
				rid: string;
				t: SubscriptionType;
				tmid?: string;
				message?: TMessageModel;
				name?: string;
				fname?: string;
				prid?: string;
				room?: TSubscriptionModel | { rid: string; t: string; name?: string; fname?: string; prid?: string };
				jumpToMessageId?: string;
				jumpToThreadId?: string;
				roomUserId?: string | null;
				usedCannedResponse?: string;
				status?: string;
		  }
		| undefined; // Navigates back to RoomView already on stack
};